requirements.txt
asgiref==3.8.1 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
channels==4.1.0 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
click==8.1.7 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
colorama==0.4.6 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0" and platform_system == "Windows"
django==5.1.3 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
h11==0.14.0 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
python-decouple==3.8 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
sqlparse==0.5.1 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
typing-extensions==4.12.2 ; python_full_version >= "3.10.12" and python_version < "3.11"
tzdata==2024.2 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0" and sys_platform == "win32"
uvicorn==0.32.0 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
whitenoise==6.8.2 ; python_full_version >= "3.10.12" and python_full_version < "4.0.0"
